chore: refactored unittest for comparison of EMPTY with custom objects
